+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Current
 
-- Add `fetch_pair_details` to get info on Uniswap v2 pairs
+- Feature: generic price oracle implementation with configurable price function
+- Feature: time weighted average price (TWAP) price function for price oracle
+- Feature: price oracle implementation for Uniswap v2 pools
+- Feature: `fetch_pair_details` to get info on Uniswap v2 pairs
 
 # 0.10.0